requirements.txt
djangorestframework
djangorestframework-authtoken